frontend: add board websocket

+++ shared/src/commonMain/kotlin/to/bnt/draw/shared/structures/Figure.kt
@@ -3,4 +3,4 @@ package to.bnt.draw.shared.structures
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Figure(var id: Int?, val drawingData: String, val color: String, val strokeWidth: Int)
+data class Figure(var id: Long?, val drawingData: String, val color: String, val strokeWidth: Int)
